Set up directory structure, add Podfile, and add a linter.
