+++ .swiftlint.yml
@@ -1,0 +1,5 @@
+disabled_rules:
+  - line_length
+
+excluded:
+- Pods
