+++ Podfile
@@ -1,0 +1,23 @@
+platform :ios, '11.1'
+
+target 'SocialTodo' do
+  use_frameworks!
+
+  pod 'FacebookCore'
+  pod 'FacebookLogin'
+  pod 'SwiftLint'
+end
+
+target 'SocialTodoTests' do
+  use_frameworks!
+
+  pod 'FacebookCore'
+  pod 'FacebookLogin'
+end
+
+target 'SocialTodoUITests' do
+  use_frameworks!
+
+  pod 'FacebookCore'
+  pod 'FacebookLogin'
+end
