--- SocialTodo/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  SocialTodo
-//
-//  Created by Saatvik Arya on 12/2/17.
-//  Copyright © 2017 Saatvik Arya. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-	override func viewDidLoad() {
-		super.viewDidLoad()
-		// Do any additional setup after loading the view, typically from a nib.
-	}
-
-	override func didReceiveMemoryWarning() {
-		super.didReceiveMemoryWarning()
-		// Dispose of any resources that can be recreated.
-	}
-
-
-}
-
