+++ SocialTodo/ViewControllers/ViewController.swift
@@ -1,0 +1,23 @@
+//
+//  ViewController.swift
+//  SocialTodo
+//
+//  Created by Saatvik Arya on 12/2/17.
+//  Copyright © 2017 Saatvik Arya. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+  override func viewDidLoad() {
+    super.viewDidLoad()
+
+  }
+
+  override func didReceiveMemoryWarning() {
+    super.didReceiveMemoryWarning()
+
+  }
+
+}
