+++ SocialTodoTests/SocialTodoTests.swift
@@ -1,0 +1,31 @@
+//
+//  SocialTodoTests.swift
+//  SocialTodoTests
+//
+//  Created by Brannen Hall on 17-12-02.
+//  Copyright © 2017 Saatvik Arya. All rights reserved.
+//
+
+import XCTest
+
+class SocialTodoTests: XCTestCase {
+
+  override func setUp() {
+    super.setUp()
+  }
+
+  override func tearDown() {
+    super.tearDown()
+  }
+
+  func testExample() {
+
+  }
+
+  func testPerformanceExample() {
+    self.measure {
+
+    }
+  }
+
+}
