+++ SocialTodoUITests/SocialTodoUITests.swift
@@ -1,0 +1,29 @@
+//
+//  SocialTodoUITests.swift
+//  SocialTodoUITests
+//
+//  Created by Brannen Hall on 17-12-02.
+//  Copyright © 2017 Saatvik Arya. All rights reserved.
+//
+
+import XCTest
+
+class SocialTodoUITests: XCTestCase {
+
+  override func setUp() {
+    super.setUp()
+
+    continueAfterFailure = false
+
+    XCUIApplication().launch()
+  }
+
+  override func tearDown() {
+    super.tearDown()
+  }
+
+  func testExample() {
+
+  }
+
+}
